+++ tests/DatumIngest.Tests/Execution/ParallelOperatorTests.cs
@@ -317,7 +317,7 @@ public sealed class ParallelOperatorTests
 
         Assert.Single(results);
         Assert.Equal(60f, results[0]["SUM(price)"].AsFloat32());
-        Assert.Equal(20f, results[0]["AVG(price)"].AsFloat32());
+        Assert.Equal(20.0, results[0]["AVG(price)"].AsFloat64());
     }
 
     /// <summary>
